Refactored base command class
