+++ command/command.cpp
@@ -1,5 +1,5 @@
 #include "command.h"
 
-command::command(std::istream *inputStream, std::ostream *outputStream): input(inputStream), output(outputStream){
+command::command(std::istream &inputStream, std::ostream &outputStream): input(&inputStream), output(&outputStream){
 
 }
