+++ command/command.h
@@ -9,7 +9,7 @@ protected:
     std::istream *input;
     std::ostream *output;
 public:
-    command(std::istream* inputStream, std::ostream* outputStream);
+    command(std::istream &inputStream, std::ostream &outputStream);
     virtual int Execute() = 0;
 };
