+++ command/hashing/base64.cpp
